Update Archived page to include FAQs and videos and Documents to have GitHub links
