+++ src/app/(resources)/docs/page.tsx
