--- src/app/(resources)/faqs/page.tsx
@@ -1,12 +1,0 @@
-import FAQHomeComponent from '@/components/resources/FAQHome'
-import CommunicationFab from '@/components/shared/CommunicationFab'
-const FAQHome = () => {
-  return (
-    <>
-      <FAQHomeComponent />
-      <CommunicationFab />
-    </>
-  )
-}
-
-export default FAQHome
