--- src/components/resources/FAQHome.cy.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import FAQHome from './FAQHome'
-
-describe('<FAQHome />', () => {
-  it('renders', () => {
-    cy.mount(<FAQHome />)
-  })
-})
