+++ src/components/shared/Ankle.tsx
@@ -16,19 +16,11 @@ declare global {
 }
 
 export default function Ankle() {
-  const ClickFAQAnkle = () => {
-    if (typeof window.gtag === 'function') {
-      window.gtag('event', 'Go to FAQs', {
-        event_category: 'Link',
-        event_label: 'Go to FAQs thru ankle',
-      })
-    }
-  }
   const ClickDocumentationAnkle = () => {
     if (typeof window.gtag === 'function') {
-      window.gtag('event', 'Go to Docs & Videos', {
+      window.gtag('event', 'Go to Docs', {
         event_category: 'Link',
-        event_label: 'Go to Docs & Videos thru ankle',
+        event_label: 'Go to Docs thru ankle',
       })
     }
   }
@@ -44,18 +36,13 @@ export default function Ankle() {
     <Box bgcolor="#122953" boxShadow={'0px -4px 20px 20px rgb(0, 0, 0, 0.15)'} pt={4} pb={4}>
       <Container>
         <Typography variant="h5" color="#fff" component={'h2'}>
-          <strong>Cant find what your looking for?</strong>
+          <strong>Can't find what your looking for?</strong>
         </Typography>
         <Typography variant="body1" color="#fff">
           Please checkout our{' '}
-          <Link onClick={ClickFAQAnkle} style={{ color: palette.white }} href="/faqs" passHref>
-            FAQs
-          </Link>
-          {', '}
-          <Link onClick={ClickDocumentationAnkle} style={{ color: palette.white }} href="/docs-and-vids" passHref>
-            Documentation &amp; Videos
-          </Link>
-          {', '}
+          <Link onClick={ClickDocumentationAnkle} style={{ color: palette.white }} href="/docs" passHref>
+            Documentation
+          </Link>{' '}
           or{' '}
           <Link onClick={ClickArchivedAnkle} style={{ color: palette.white }} href="/archived" passHref>
             Archived
